feat(dashboard): Merging #93 - The codebase of @Huskydog9988 will be used as the new codebase of this project
* Adding a login system

* Added a login system

* Changed over to ejs

Not everything is moved yet, but the core ideas are.

* Shift in the base api to a class based system

Changed the api from run() to CLASS.run() basicly and did some stuff with ejs.

* Many changes

Added the very basic foundation for a sql logging system.

Added live stats to the site for guilds being joined.

Added error pages.

Added some test commands to the test.js file for the bot.

* Major changes in appearance and functionality

First, imporved the ui of the dashbaord by unifying most of the top row to look the same. Next, added db items to gitignore and npmignore. Semi improved some of the docs. Added a feature to generate a redirect URI, but that requires some work. Plan on adding resrouce logging, started it with "resoruceCheck.js" but for the most part that should be sidelined. All the files under the folder "sql" should have the same treatment as "resourceCheck". Added an AdminLTE skin called "midnight". Havn't implemented as per a plan to have a light and dark theme. Removed the last of "/db" with the "placeholder.txt" as the user will just specify where the db file goes as they will pass the db object to the dashboard.